+++ Backend/GlobalSuppressions.cs
@@ -1,0 +1,5 @@
+﻿// Copyright (C) 2024 Claudia Wagner
+
+using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Style", "IDE0290:Use primary constructor", Justification = "Imho matter of taste", Scope = "module")]