+++ Backend/Model/Configuration.cs
@@ -1,4 +1,6 @@
-﻿using Common;
+﻿// Copyright (C) 2024 Claudia Wagner, Daniel Kuster
+
+using Common;
 using Microsoft.Extensions.Configuration;
 using System.Text.Json.Serialization;
 
@@ -15,8 +17,13 @@ namespace Backend.Model {
 
         public int PortLogcat { get; set; }
 
+        public bool AllowAnonymousLogs { get; set; }
+
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public LogTimeFormat LogTimeFormat { get; set; }
+
+        [JsonConverter(typeof(JsonStringEnumConverter))]
+        public ApplicationFormat ApplicationFormat { get; set; }
     }
 
     public static class ConfigurationExtensions {