+++ Common/ApplicationFormat.cs
@@ -1,0 +1,10 @@
+﻿// Copyright (C) 2024 Claudia Wagner
+
+namespace Common {
+
+    public enum ApplicationFormat {
+        None,
+        DoNotChange,
+        Consolidate
+    }
+}