+++ Common/Constants.cs
@@ -1,11 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿// Copyright (C) 2024 Claudia Wagner, Daniel Kuster
+
+using System.Text.RegularExpressions;
 
 namespace Common {
-    public static class Constants {
+
+    public static partial class Constants {
 
         public const int DEFAULT_MAX_NUMBER_OF_LOGS_PER_LEVEL = 1000;
 
@@ -15,5 +14,8 @@ namespace Common {
         public const string NAMESPACE_LOGCAT = "Logcat";
         public const string NAMESPACE_GLOBAL = "Global (namespace)";
         public const string APPLICATION_GLOBAL = "Global (application)";
+
+        [GeneratedRegex(@"^(?<app>.+)\((?<pid>[^)]+)\)\s*$", RegexOptions.Compiled)]
+        public static partial Regex RegexLog4jApp();
     }
 }